2016-12-10 Saturday Morning I spent for several hours on it, by using Clion's debug func, finally I run it successfully.

+++ Stack/constant.h
@@ -19,6 +19,6 @@
 #define OVERFLOW -2
 
 typedef int Status;
-typedef int ElemType;
+typedef double ElemType;
 
 #endif //STACK_CONSTANT_H
